# tools/audit/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TraderCopilot Backend",
    "version": "2.0.0"
  },
  "paths": {
...
    "/advisor/": {
      "post": {
        "tags": [
          "Advisor",
          "advisor"
        ],
        "summary": "Analyze Advisor",
        "description": "Analiza una posici\u00f3n abierta y sugiere alternativas.\nVersi\u00f3n local determinista (sin LLM).\nEnforces Entitlements.",
        "operationId": "analyze_advisor_advisor__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvisorReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/notify/telegram": {
      "post": {
        "summary": "Notify Telegram",
        "description": "Env\u00eda una notificaci\u00f3n a Telegram (simulado por ahora).",
        "operationId": "notify_telegram_notify_telegram_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TelegramMsg"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/strategies/marketplace/{id}/toggle": {
      "patch": {
        "summary": "Toggle Strategy Active",
        "description": "Toggles the is_active state of a strategy persona.\nSupports both Custom (user_strategies.json) and System (system_overrides.json) personas.",
        "operationId": "toggle_strategy_active_strategies_marketplace__id__toggle_patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AdvisorReq": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Token base: eth|btc|sol|xau"
          },
          "direction": {
            "type": "string",
            "enum": [
              "long",
              "short"
            ],
            "title": "Direction",
            "description": "Direcci\u00f3n de la posici\u00f3n"
          },
          "entry": {
            "type": "number",
            "title": "Entry",
            "description": "Precio de entrada de la posici\u00f3n"
          },
          "tp": {
            "type": "number",
            "title": "Tp",
            "description": "Take profit previsto"
          },
          "sl": {
            "type": "number",
            "title": "Sl",
            "description": "Stop loss previsto"
          },
          "size_quote": {
            "type": "number",
            "title": "Size Quote",
            "description": "Tama\u00f1o de la posici\u00f3n en moneda de cotizaci\u00f3n (ej. USDT)"
          }
        },
        "type": "object",
        "required": [
          "token",
          "direction",
          "entry",
          "tp",
          "sl",
          "size_quote"
        ],
        "title": "AdvisorReq"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TelegramMsg": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "TelegramMsg"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/token"
          }
        }
      }
    }
  }
}